feat: formalize integration test tiers with directories and markers
Organize integration tests into subdirectories by environment requirements
so developers can run any tier by right-clicking the directory in their IDE:

- host/    — needs Linux kernel only (IP_RECVERR, filesystem), runs in CI
- network/ — needs outbound connectivity + dig
- podman/  — needs podman + nft (+ internet)

Each directory has a matching Makefile target (test-host, test-network,
test-podman) plus test-integration for all tiers. Fine-grained pytest
markers (needs_host_features, needs_podman, needs_internet) are applied
to each test class for cross-directory selection.

Split test_config_detect.py and test_profiles_live.py across tiers where
classes had different requirements. Add host-only integration tests to CI.

Closes #23
