+++ .github/workflows/integration.yml
@@ -41,4 +41,4 @@ jobs:
         run: sudo apt-get update && sudo apt-get install -y nftables podman
 
       - name: Run integration tests
-        run: poetry run pytest tests/integration/ -v -m integration
+        run: poetry run pytest tests/integration/ -v
