+++ tests/integration/host/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-FileCopyrightText: 2026 Jiri Vyskocil
+# SPDX-License-Identifier: Apache-2.0
